pyproject.toml
[project]
name = "salai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "py3dmol>=2.5.0",
    "pyqt5>=5.15.11",
    "rdkit>=2025.3.3",
    "streamlit>=1.46.0",
    "torch>=2.7.1",
    "torch-geometric>=2.6.1",
]
